Use Guzzle 4, 5, and 6 only
Guzzle 3 has been EOL for 5 years. 4 has been EOL for 3 years (though it
had a release ~21 months ago). Guzzle 5 is still being maintained, but
is considered legacy to the PSR-7-based Guzzle 6. Versions 4 and 5 have
the same API for our purposes, so we can support those two reasonably,
and add support for version 6 (which, while we had code to allow it, it
never was actually tested, and, in fact, did not work without
deprecation notices).