Make process checks work better
As reported in #82, on some operating systems, `ps -ef` will not work as
intended if the window is too small to show the entire process name. The
solution to this is to enable unlimited width output, which causes each
line to wrap (`-ww`).

This patch also does method extraction refactors to simplify the logic.